Update sensor.py
When a SerialException or OSError is raised, the module attempts to reconnect immediately, which it often fails to do as these exceptions don't usually solve themselves. This causes the module to continuously spam the logs, resulting in GBs of data after a few days, and increased cpu usage. 

Adding a return statement so it logs the issue, then stops reconnecting.